% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRdelayBD-MNRM.R
\name{SEIRdelay_BD_MNRM_C}
\alias{SEIRdelay_BD_MNRM_C}
\title{SEIR model with fixed delay and demographics via Modified Next Reaction Method (\strong{C} version)}
\usage{
SEIRdelay_BD_MNRM_C(
  tmax,
  S0,
  I0,
  beta,
  nu,
  tau,
  mu,
  verbose = T,
  maxsize = 1e+06
)
}
\arguments{
\item{tmax}{maximum time of simulation}

\item{S0}{initial number of susceptibles}

\item{I0}{initial number of infectives}

\item{beta}{effective contact rate}

\item{nu}{duration of infectiousness}

\item{tau}{duration of latent period fixed delay}

\item{mu}{mortality rate}

\item{verbose}{print diagnostic information}

\item{maxsize}{maximum size of output matrix; simulation will return early if this is exceeded}
}
\description{
Using the Modified Next Reaction Method (MNRM) for delays, simulate a SEIR
model where the time spent in the E compartment is a deterministic period of
length \code{tau}. There are additional events for births in to S and deaths from
all compartments according to rate \code{mu} such that the total population is at
its stationary distribution.
}
\details{
The MNRM is described in: Anderson, D. F. (2007). \emph{A modified next reaction method for simulating chemical systems with time dependent propensities and delays.} Journal of Chemical Physics, 127(21). \url{https://doi.org/10.1063/1.2799998}
}
\examples{
 \dontrun{
   out <- SEIRdelay_BD_MNRM_C(tmax = 1e3,S0 = 1000,I0 = 5,beta = 0.001,nu = 1/5,tau = 10,mu = 1/50,verbose = TRUE)
   outd <- discretize(out)
   matplot(outd[,-1],type="l",col=c("blue","orange","red","purple"),lwd=1.5,lty=1,ylab="Count",xlab="Days")
   legend(x = "topleft",pch=rep(16,4),col=c("blue","orange","red","purple"),legend=c("S","E","I","R"))
}
}
