% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRdelay-MNRM.R
\name{SEIRdelay_MNRM}
\alias{SEIRdelay_MNRM}
\title{SEIR model with fixed delay via Modified Next Reaction Method}
\usage{
SEIRdelay_MNRM(tmax, S0, I0, beta, nu, tau, verbose = T, maxsize = 1e+06)
}
\arguments{
\item{tau}{duration of latent period fixed delay}

\item{verbose}{print diagnostic information}
}
\description{
SEIR model with fixed delay via Modified Next Reaction Method
}
