% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRdelay-MNRM.R
\name{SEIRdelay_MNRM_C}
\alias{SEIRdelay_MNRM_C}
\title{SEIR model with fixed delay via Modified Next Reaction Method (\strong{C} version)}
\usage{
SEIRdelay_MNRM_C(tmax, S0, I0, beta, nu, tau, verbose = T, maxsize = 1e+06)
}
\arguments{
\item{tau}{duration of latent period fixed delay}

\item{verbose}{print diagnostic information}
}
\description{
SEIR model with fixed delay via Modified Next Reaction Method (\strong{C} version)
}
\examples{
 \dontrun{
   out <- SEIRdelay_MNRM_C(tmax = 1e3,S0 = 1000,I0 = 5,beta = 0.001,nu = 1/5,tau = 10,verbose = TRUE)
   outd <- discretise(out)
   matplot(outd[,-1],type="l",col=c("blue","orange","red","purple"),lwd=1.5,lty=1,ylab="Count",xlab="Days")
   legend(x = "topleft",pch=rep(16,4),col=c("blue","orange","red","purple"),legend=c("S","E","I","R"))
}
}
