% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRdelay-MNRM.R
\name{SEIRdelay_MNRM}
\alias{SEIRdelay_MNRM}
\title{SEIR model with fixed delay via Modified Next Reaction Method (\strong{R} version)}
\usage{
SEIRdelay_MNRM(tmax, S0, I0, beta, nu, tau, verbose = T, maxsize = 1e+06)
}
\arguments{
\item{tmax}{maximum time of simulation}

\item{S0}{initial number of susceptibles}

\item{I0}{initial number of infectives}

\item{beta}{effective contact rate}

\item{nu}{duration of infectiousness}

\item{tau}{duration of latent period fixed delay}

\item{verbose}{print diagnostic information}

\item{maxsize}{maximum size of output matrix; simulation will return early if this is exceeded}
}
\description{
Using the Modified Next Reaction Method (MNRM) for delays, simulate a SEIR
model where the time spent in the E compartment is a deterministic period of
length \code{tau}.
}
\details{
The MNRM is described in: Anderson, D. F. (2007). \emph{A modified next reaction method for simulating chemical systems with time dependent propensities and delays.} Journal of Chemical Physics, 127(21). \url{https://doi.org/10.1063/1.2799998}
}
\examples{
 \dontrun{
   out <- SEIRdelay_MNRM(tmax = 1e3,S0 = 1000,I0 = 5,beta = 0.001,nu = 1/5,tau = 10,verbose = TRUE)
   outd <- discretise(out)
   matplot(outd[,-1],type="l",col=c("blue","orange","red","purple"),lwd=1.5,lty=1,ylab="Count",xlab="Days")
   legend(x = "topleft",pch=rep(16,4),col=c("blue","orange","red","purple"),legend=c("S","E","I","R"))
}
}
